feat:分享链接接口verifyShare完成

+++ src/main/java/com/liteisle/common/domain/response/ShareInfoResp.java
@@ -8,5 +8,5 @@ public class ShareInfoResp {
     private ItemType itemType;  // "file" 或 "folder"
     private String itemName;
     private Long itemSize;    // 如果是文件，为文件大小；如果是文件夹，为文件夹总大小
-    private Integer totalFiles;  // 如果是文件夹，为包含的文件总数
+    private Long totalFiles;  // 如果是文件夹，为包含的文件总数
 }
